+++ crates/header-translator/src/lib.rs
@@ -16,6 +16,7 @@ mod config;
 mod context;
 mod display_helper;
 mod expr;
+mod fn_utils;
 mod global_analysis;
 mod id;
 mod library;