+++ framework-crates/objc2-metal/tests/runs_with_core_graphics.rs
@@ -1,5 +1,4 @@
 #![cfg(feature = "MTLDevice")]
-use objc2::rc::Retained;
 use objc2_metal::MTLCreateSystemDefaultDevice;
 
 #[link(name = "CoreGraphics", kind = "framework")]
@@ -8,11 +7,11 @@ extern "C" {}
 #[test]
 #[ignore = "doesn't work in CI"]
 fn test_create_default() {
-    let _ = unsafe { Retained::from_raw(MTLCreateSystemDefaultDevice()).unwrap() };
+    let _ = MTLCreateSystemDefaultDevice();
 }
 
 #[test]
 #[cfg(target_os = "macos")]
 fn get_all() {
-    let _ = unsafe { Retained::from_raw(objc2_metal::MTLCopyAllDevices().as_ptr()).unwrap() };
+    let _ = objc2_metal::MTLCopyAllDevices();
 }